* automatically add destination filename in copy/move dialog only when source is one file + add /*.* to destination of copy/move dialog when destination is directory (important for implementation of copy/move operations) * fix of remove directories in move operation

+++ FileOperation/CopyMove.cpp
@@ -306,14 +306,16 @@ void cCopyMove::run()
 																	qi64TotalValue += qfilSources.at(iI).size();
 																	emit SetTotalValue(qi64TotalValue);
 																} // if else
-																qdDir.rmdir(qfilSources.at(iI).path());
+																if (qsSourcePath != qfilSources.at(iI).path()) {
+																	qdDir.rmdir(qfilSources.at(iI).path());
+																} // if
 																break;
 			} // switch
 		} // if else
 	} // for
 
 	// remove source directory when moving files
-	if (cFileRoutine::MoveOperation) {
+	if (eoOperation == cFileRoutine::MoveOperation) {
 		qdDir.rmdir(qfilSource.at(0).filePath());
 	} // if
